First Laravel app import

+++ database/migrations/2022_06_28_201257_create_data_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->integer('year');
-            $table->integer('capacity');
-            $table->integer('production');
+            $table->float('capacity');
+            $table->float('production');
             $table->timestamps();
         });
     }
